add a convenience envtemplate docker image (#5685)
The idea here is to simplify some open-source dockerfiles that need envtemplate for doing environment-based runtime file templating.

So we add a private dockerfile for envtemplate, and in a couple of public Dockerfiles we pull the binary from here instead of building it. This makes the builds simpler, as we don't have to install git and golang just to get a simple binary.

This will also make #5670 a little easier to land.

fbshipit-source-id: 7acd700